# gohandlers/go/model_email_verify.go
/*
 * Collaborative shopping
 *
 * Collaboratively go shopping. Create shopping lists, share them with others, bring their needs along 
 *
 * API version: 1.0.0
 * Contact: collaborative-shopping@beimir.net
 * Generated by: Swagger Codegen (https://github.com/swagger-api/swagger-codegen.git)
 */
package swagger

type EmailVerify struct {

	Email string `json:"email"`
}