# gohandlers/go/model_password_reset.go
/*
 * Collaborative shopping
 *
 * Collaboratively go shopping. Create shopping lists, share them with others, bring their needs along 
 *
 * API version: 1.0.0
 * Contact: collaborative-shopping@beimir.net
 * Generated by: Swagger Codegen (https://github.com/swagger-api/swagger-codegen.git)
 */
package swagger

type PasswordReset struct {

	Password string `json:"password"`

	PasswordConfirm string `json:"password_confirm"`
}